+++ win32/mssql/test/test_mssql_transaction.cpp
@@ -30,7 +30,7 @@ TEST( mssql_transaction, should_commit_on_destruction )
    {
       ds::db::statement results_test = test_db( num_rows );
 
-      int count = results_test.execute();
+      uint64_t count = results_test.execute();
 
       EXPECT_EQ( 2, count );
    }
@@ -68,7 +68,7 @@ TEST( mssql_transaction, should_rollback_on_exception )
    {
       ds::db::statement results_test = test_db( num_rows );
 
-      int count = results_test.execute();
+      uint64_t count = results_test.execute();
 
       EXPECT_EQ( 0, count );
    }
@@ -162,7 +162,7 @@ TEST( mssql_savepoint, should_release_on_destruction )
    {
       ds::db::statement results_test = test_db( num_rows );
 
-      int count = results_test.execute();
+      uint64_t count = results_test.execute();
 
       EXPECT_EQ( 2, count );
    }
@@ -205,7 +205,7 @@ TEST( mssql_savepoint, should_rollback_on_exception )
    {
       ds::db::statement results_test = test_db( num_rows );
 
-      int count = results_test.execute();
+      uint64_t count = results_test.execute();
 
       EXPECT_EQ( 1, count );
    }
